Move synth components to a separate package, delete old BA
