+++ csb/src/main/scala/edu/msstate/dasi/csb/synth/GraphSynth.scala
@@ -1,7 +1,8 @@
-package edu.msstate.dasi.csb
+package edu.msstate.dasi.csb.synth
 
 import edu.msstate.dasi.csb.distributions.DataDistributions
 import edu.msstate.dasi.csb.model.{EdgeData, VertexData}
+import edu.msstate.dasi.csb.{Util, sc}
 import org.apache.spark.graphx.Graph
 
 /**
